pyproject.toml
[tool.poetry]
name = "dzr-tui"
version = "0.1.0"
description = ""
authors = ["Vinicius Costa"]
readme = "README.md"
packages = [{include = "dzr_tui"}]

[tool.poetry.dependencies]
python = "^3.8"
deezer-py = "^1.3.7"
pytermgui = "^7.2.0"
pyyaml = "^6.0"
python-mpv = "^1.0.1"
click = "^8.1.3"
deemix = "^3.6.6"
python-mpd2 = "^3.0.5"


[tool.poetry.group.dev.dependencies]
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dzr = 'dzr_tui:__main__.main'